+++ Source/HexNoise/Public/FPerlinNoiseGenerator1D.h
@@ -20,7 +20,7 @@ public:
 	* @param X	The X coordinate
 	* @returns	The noise value
 	*/
-	float GetNoise(float X);
+	double GetNoise(double X);
 
 	/** Changes the settings used by the generator
 	* @param Settings	Pointer to the new settings
@@ -34,29 +34,29 @@ private:
 	* @param X	The X coordinate
 	* @returns	The noise value
 	*/
-	float RawNoise(int32 X);
+	double RawNoise(int32 X);
 
 	/** Generates interpolated and (if desired) smoothed noise for a specified x coordinate, used for interpolations requiring three params;
 	* @param X	The X coordinate
 	* @returns	The noise value
 	*/
-	float InterpolatedNoise(float X);
+	double InterpolatedNoise(double X);
 
 	/** Generates interpolated and (if desired) smoothed noise for a specified x coordinate, used for interpolations requiring five params;
 	* @param X	The X coordinate
 	* @returns	The noise value
 	*/
-	float AdvancedInterpolatedNoise(float X);
+	double AdvancedInterpolatedNoise(double X);
 
 	/** Generates smoothed noise for a specified x coordinate
 	* @param X	The X coordinate
 	* @returns	The noise value
 	*/
-	float SmoothNoise(float X);
+	double SmoothNoise(double X);
 
 	/** Pointer to the function fitting to the selected interpolation method
 	* e.g when selected "Lerp" this will point to a function which only calculates data for V1 and V2
 	* but when "Cubic" is selected this will point to a function which also calculates data for V0 and V3
 	*/
-	float (FPerlinNoiseGenerator1D::*InterpHub)(float);
+	double (FPerlinNoiseGenerator1D::*InterpHub)(double);
 };
